duplicate check from operations to cdm_mapper
