+++ tests/test_operations.py
@@ -2,54 +2,10 @@ from __future__ import annotations
 
 import pytest  # noqa
 
-from cdm_reader_mapper.cdm_mapper import read_tables
-from cdm_reader_mapper.operations import (
-    corrections,
-    duplicates,
-    inspect,
-    replace,
-    select,
-)
+from cdm_reader_mapper.operations import corrections, inspect, replace, select
 
 from ._data import data_df, data_pa, mask_df, mask_pa
-from ._results import correction_df, result_data, table_df
-
-
-def _manipulate_header(df):
-    df.loc[5] = df.loc[4]
-    df.loc[5, "report_id"] = "ICOADS-302-N688EY"
-    df.loc[5, "source_record_id"] = "N688EY"
-    df.loc[6] = df.loc[4]
-    df.loc[6, "latitude"] = -65.80
-    df.loc[6, "longitude"] = 21.20
-    df.loc[7] = df.loc[0]
-    df.loc[7, "source_record_id"] = "N688DP"
-    df.loc[8] = df.loc[1]
-    df.loc[8, "report_timestamp"] = "2022-02-02 00:00:00"
-    df.loc[9] = df.loc[2]
-    df.loc[9, "report_id"] = "ICOADS-302-N688DW"
-    df.loc[10] = df.loc[3]
-    df.loc[10, "latitude"] = 9.10
-    df.loc[10, "longitude"] = 68.00
-    return df
-
-
-def _manipulate_observations(df):
-    df.loc[3] = df.loc[0]
-    df.loc[3, "observation_id"] = "ICOADS-302-N688DP-SST"
-    df.loc[3, "source_id"] = "ICOADS-3-0-2T-103-792-2022-3"
-    df.loc[4] = df.loc[0]
-    df.loc[4, "latitude"] = -71.30
-    df.loc[5] = df.loc[1]
-    df.loc[5, "source_id"] = "ICOADS-3-0-2T-103-792-2022-3"
-    df.loc[6] = df.loc[1]
-    df.loc[6, "date_time"] = "2022-02-02 00:00:00"
-    df.loc[7] = df.loc[2]
-    df.loc[7, "observation_id"] = "ICOADS-302-N688DW-SST"
-    df.loc[8] = df.loc[2]
-    df.loc[8, "latitude"] = 73.20
-    df.loc[8, "longitude"] = 34.00
-    return df
+from ._results import correction_df, table_df
 
 
 def test_select_true_pandas():
@@ -113,29 +69,3 @@ def test_replace():
         pivot_c="report_id",
         rep_c=["primary_station_id", "primary_station_id.isChange"],
     )
-
-
-def test_duplicates_header():
-    expected_data = result_data.expected_103_792
-    data_path = expected_data.get("cdm_table")
-    df = read_tables(
-        data_path,
-        tb_id="103-792*",
-        cdm_subset="header",
-    )
-    df = _manipulate_header(df)
-    DupDetect = duplicates.duplicate_check(df)
-    return DupDetect.remove_duplicates()
-
-
-def test_duplicates_observations():
-    expected_data = result_data.expected_103_792
-    data_path = expected_data.get("cdm_table")
-    df = read_tables(
-        data_path,
-        tb_id="103-792*",
-        cdm_subset="observations-sst",
-    )
-    df = _manipulate_observations(df)
-    DupDetect = duplicates.duplicate_check(df, cdm_name="observations")
-    return DupDetect.remove_duplicates()
